Animal observation records from point surveys

Time-based point observations of species abundance at sampling points within six towns (areas) in Singapore: Punggol (PG), Queenstown (QT), Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).

Usage

data(animal_observations)

Format

Dataframe

Source

Development of a Biodiversity Index for Residential Towns using Biodiversity Field Surveys. 14 June 2016, 6.5 years, Ministry of National Development Research Fund (MNDRF) Grant, $1,286,979.79. Awarded to the National University of Singapore (host institution) and the Singapore Housing Development Board.

Details

The sampling points (point_id) were randomly generated within each town (area). Four types of animal surveys were conducted at sampling points, corresponding to four animal groups (column taxon): Birds (Aves), Butterflies (Lepidoptera), Odonates (Odonata) and Amphibians (Amphibia). In addition to the name of the species and abundance (number observed) at the particular time point, the names of the family and genus are also made available.

Each sampling period stretched across a year-long duration, where sampling points were sampled once every two months (cycle adds up to 6 per period). Some points were sampled again in the second period (column resampled is TRUE). All surveys took place between the years 2016 to 2022.

See also

Examples

head(animal_observations)
<<<<<<< HEAD
#> # A tibble: 6 × 13
#>   survey_id            point_id area  period cycle resampled start_time         
#>   <chr>                <chr>    <chr>  <dbl> <dbl> <lgl>     <dttm>             
#> 1 1 QTNa14a_P 1 Odona… QTNa14a… QT         1     1 NA        2016-08-04 14:00:00
#> 2 1 QTNa14a_P 1 Odona… QTNa14a… QT         1     1 NA        2016-08-04 14:00:00
#> 3 1 QTNa14a_P 1 Odona… QTNa14a… QT         1     1 NA        2016-08-04 14:00:00
#> 4 1 QTNa14a_P 1 Odona… QTNa14a… QT         1     1 NA        2016-08-04 14:00:00
#> 5 1 QTNa14a_P 1 Odona… QTNa14a… QT         1     1 NA        2016-08-04 14:00:00
#> 6 1 QTNb1a_P 1 Odonata QTNb1a_P QT         1     1 NA        2016-08-04 14:44:00
#> # … with 6 more variables: time <dttm>, taxon <chr>, species <chr>,
=======
#> # A tibble: 6 x 13
#>   survey_id            point_id area  period cycle resampled start_time         
#>   <chr>                <chr>    <chr>  <dbl> <dbl> <lgl>     <dttm>             
#> 1 1 QTNa14a_P 1 Odona~ QTNa14a~ QT         1     1 NA        2016-08-04 14:00:00
#> 2 1 QTNa14a_P 1 Odona~ QTNa14a~ QT         1     1 NA        2016-08-04 14:00:00
#> 3 1 QTNa14a_P 1 Odona~ QTNa14a~ QT         1     1 NA        2016-08-04 14:00:00
#> 4 1 QTNa14a_P 1 Odona~ QTNa14a~ QT         1     1 NA        2016-08-04 14:00:00
#> 5 1 QTNa14a_P 1 Odona~ QTNa14a~ QT         1     1 NA        2016-08-04 14:00:00
#> 6 1 QTNb1a_P 1 Odonata QTNb1a_P QT         1     1 NA        2016-08-04 14:44:00
#> # ... with 6 more variables: time <dttm>, taxon <chr>, species <chr>,
>>>>>>> 6b090ae (regenerated website)
#> #   family <chr>, genus <chr>, abundance <dbl>