Calculate metrics based on manually generated data at point locations

Calculate metrics from manually generated vector data, at specific point locations and buffer radii. Currently supports vector data of buildings (polygons), roads (lines), trees (points), shrubs (polygons), turf (polygons), natural vegetation (polygons), and water (polygons).

Usage

calc_manual(
  vector,
  name = NULL,
  points,
  buffer_sizes,
  plant_species = "species",
  building_levels = "levels",
  road_lanes = "lanes"
)

Arguments

vector

sf dataframe containing geometric features (points, polygons or lines) to be summarised. Coordinate reference system should be similar to points.

name

Specify either "buildings", "roads", "trees", "shrubs', "turf", "natveg", or "water". Used to represent the type of data in vector, each to be processed differently.

points

Point locations (sf object) to calculate the metrics.

buffer_sizes

Radius of circle (in mapunits) for each point location; metrics will be calculated within the buffer area.

plant_species

Column name in vector for plant species names, if input vector is vegetation (i.e., argument name= "trees" or "shrubs"). Defaults to "species". Column data should be of type character.

building_levels

Column name in vector for the number of building levels, when name="buildings". Defaults to "levels". Column data should be numeric.

road_lanes

Column name in vector for the number of lanes per road line, when name="roads". Defaults to "lanes". Column data should be numeric.

Value

A list containing the features/metrics calculated for points, appended as new columns. Each element in the list corresponds to a particular buffer size.