Calculate landscape metrics at sampling points

Calculate landscape metrics at sampling points

Usage

lsm_perpoint(
  raster,
  points,
  buffer_sizes,
  class_names,
  class_values,
  landscape_name,
  point_id = "point_id",
  period = "period",
  what = NULL,
  level = NULL,
  ...
)

Arguments

raster

Classified land cover raster object to be analysed. The number of layers should correspond to the number of survey periods present in points.

points

Sampling points (sf object) for the calculation of landscape metrics.

buffer_sizes

Radius of circles for sampling points (in mapunits); landscape metrics will be calculated within the buffer area.

class_names

Vector (character) of land cover class names to be used to identify the corresponding integer values in class_values.

class_values

Vector of (integer) values of interest within raster. Should not include the value 0.

landscape_name

Character string for name of landscape. Used to label landscape-level metrics.

point_id

Column name of the sampling point id within the points sf.

period

Column name of the survey period within the points sf. The column should contain integers that correspond to the number of layers in raster.

what

Argument passed to landscapemetrics::calculate_lsm(). Currently only supports class- and landscape-level metrics

level

Argument passed to landscapemetrics::calculate_lsm(). Currently only supports class- and landscape-level metrics.

...

Additional arguments (e.g. type=) passed to landscapemetrics::calculate_lsm().

Value

A list containing the landscape metrics across the input buffer_sizes. Each element in the list contains results for each point_id and period, for a particular buffer size.