Sampling points with landscape data summarised at multiple buffer radii

Examples of landscape data summarised at point locations where animal surveys were conducted. These sampling points are within the six towns (areas) in Singapore, during two sampling periods. The six towns were Punggol (PG), Queenstown (QT), Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).

Usage

data(points_landscape)

Format

Dataframe containing landscape predictors (columns).

Source

Development of a Biodiversity Index for Residential Towns using Biodiversity Field Surveys. 14 June 2016, 6.5 years, Ministry of National Development Research Fund (MNDRF) Grant. Awarded to the National University of Singapore (host institution) and the Singapore Housing & Development Board.

Details

Data is in 'wide format', i.e. has duplicate rows for the unique identifier of points (column point_id), owing to the presence of multiple buffer radii per point (column radius_m, in metres). Not all landscape predictors are summarised at all buffer radii. Column names for landscape predictors are prefixed with the following:

man_: manually mapped on-site

osm_: derived from OpenStreetMap data

lsm_: landscape metrics based on land cover classes

See also

Examples

head(points_landscape)
#>   point_id area period radius_m man_natveg_pland man_water_pland man_turf_pland
#> 1 TPa40c_E   TP      2       50          0.00000         0.00000      74.487333
#> 2  TPb2a_P   TP      1       50         44.79356        39.61295      15.479083
#> 3    PGNa5   PG      1       50          0.00000         0.00000      12.792271
#> 4 PGNb4a_P   PG      2       50         65.02723        17.43035       3.106433
#> 5    PGE10   PG      2      126         11.45276              NA             NA
#> 6 QTNa44_P   QT      2       50          0.00000        13.65447      39.275742
#>   man_shrub_pland man_shrub_sprich man_tree_count man_tree_sprich
#> 1       2.0436409               26             23               9
#> 2       0.8510684               25             24              10
#> 3       6.3710769               19            147               7
#> 4       0.1736969                1             13               3
#> 5              NA               NA             NA              NA
#> 6       0.3532717               25             14               5
#>   man_buildingAvgLvl man_buildingFA_ratio man_laneDensity
#> 1           8.268138            0.7823429      0.01207494
#> 2           0.000000            0.0000000      0.00000000
#> 3           9.289540            4.4620846      0.01563074
#> 4           0.000000            0.0000000      0.00000000
#> 5                 NA                   NA              NA
#> 6           1.000000            0.1035679      0.07505683