Randomly generate sampling points within a supplied area

Randomly generate sampling points at a specified density and buffer radius, within supplied polygons. An excess of sampling points can be generated, in case some of the generated points are unsuitable for surveys (e.g. within inaccessible areas). If necessary, stratified random sampling can be performed to ensure sufficient representation between sub-areas of interest. By supplying the sub_areas of interest, sampling points will be stratified between the areas outside ('normal' areas) and within these 'sub-areas'.

Usage

random_pt_gen(
  boundaries,
  area_per_pt,
  pt_radius,
  excess_modifier = 1,
  sub_areas = NULL,
  retain = NULL,
  retain_prop = NULL
)

Arguments

boundaries

Shapefile (polygon or multipolygon) of site boundaries for sampling.

area_per_pt

Area component for point density in square meters (e.g. one point every 50 hectares, area_per_pt = 500000).

pt_radius

Buffer radius of survey points in meters. Ensures that the generated points do not extend beyond the boundaries of the area of interest.

excess_modifier

Numeric value of a multiplicative modifier for excess points to be generated (i.e. a value of 1.5 = 1.5 times the expected number of points will be generated). The default value is '1'.

sub_areas

Shapefile (polygon or multipolygon) of sub-area of interest, where sampling points will be generated separately (stratified random sampling).

retain

Shapefile (point) of previous round of surveys to retain in the new set of survey points. The column type must exist, with the value of either Normal or Sub-area.

retain_prop

Value of 0 to 1 for proportion of points to retain from previous round of surveys.

Value

The sf object sampling_points containing the geographical coordinates of generated points.