Tally number of species based on specified criteria

Wrapper function to filter_obs() which filters species observations from biodiversity surveys, based on specified criteria. Subsequently tallies the number of species per survey/sampling point (similar to function specnumber() in package vegan), and extracts output data. If tallied at the level of sampling points (level = "point"), removes taxon group-level (genus/family) if all species within group are observed at the point (& round), using the function check_taxongrps(). Data on both the biodiversity surveys and their species observations (and respective abundances) must be provided.

sptally_extractor(
  observations,
  survey_ref,
  specify_town,
  specify_round,
  specify_priority,
  level,
  survey_id = "survey_id",
  town = "town",
  round = "round",
  priority = "priority",
  point_id = "point_id",
  species = "species",
  genus = "genus",
  family = "family",
  abundance = "abundance"
)

Arguments

observations
Dataframe of species observations. It should include columns for survey_id, town, round, priority, species and abundance.

survey_ref
Dataframe of all surveys conducted. Values in the column survey_id should correspond to those in the observations.

specify_town
Specify the town (character).

specify_round
Specify the survey round (character).

specify_priority
Specify the priority taxon (character).

level
Specify whether to tally by "survey" or "point" (character).

survey_id
Column name of the unique identifier for each survey in observations and survey_ref. Defaults to survey_id.

town
Column name of the town specified in observations and survey_ref. Defaults to town.

round
Column name of the sampling round specified in observations and survey_ref. Defaults to round.

priority
Column name of the priority taxon specified in observations and survey_ref. Defaults to priority.

point_id
Column name of the unique identifier for each sampling point specified in observations and survey_ref. Defaults to point_id.

species
Column name of the species specified in observations. Defaults to species.

genus
Column name of the genus specified in observations. Defaults to genus. Used to remove genus-level records in the species column, if all species within the group are observed within the filtered dataset.

family
Column name of the family specified in observations. Defaults to family. Used to remove family-level records in the species column, if all species within the group are observed within the filtered dataset.

abundance
Column name of the species abundance specified in observations. Defaults to abundance.

Value

A dataframe containing the tally of species per survey/sampling point (depends if level = "survey" or "point").