Summarise landscape variables per point based on supplied data

Summaries will be calculated for each of the supplied layers (layer_<layername>) and appended to the points data as additional columns. Ensure that all layers have a (projected) coordinate reference system similar to points.

summarise_landscape_perpt(
  points,
  uid_column = "id",
  layer_trees = NULL,
  radii_trees = 50,
  layer_shrubs = NULL,
  radii_shrubs = 50,
  species = "species",
  layer_turf = NULL,
  radii_turf = 50,
  layer_natveg = NULL,
  radii_natveg = c(50, 126),
  layer_water = NULL,
  radii_water = 50,
  layer_buildings = NULL,
  radii_buildings = 50,
  buildings_levels = "levels",
  layer_roads = NULL,
  radii_roads = c(50, 126),
  roads_lanes = "lanes"
)

Arguments

points
Sampling points (sf object) with a projected coordinate reference system.

uid_column
character. Specify unique identifier column name in points. Defaults to 'uid'. Column data should be of type character.

layer_trees
Geo-location of trees (sf points) to be summarised. Coordinate reference system should be similar to points.

radii_trees
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_trees. Default values for the model are supplied (in metres).

layer_shrubs
Vector of shrubs (sf polygons) to be summarised. Coordinate reference system should be similar to points.

radii_shrubs
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_shrubs. Default values for the model are supplied (in metres).

species
character. Specify column name for the species names within layer_trees and layer_shrubs (if provided). Defaults to 'species'. Column data should be of type character.

layer_turf
Vector of turf (sf polygons) to be summarised. Coordinate reference system should be similar to points.

radii_turf
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_turf. Default values for the model are supplied (in metres).

layer_natveg
Vector of natural vegetation (sf polygons) to be summarised. Coordinate reference system should be similar to points.

radii_natveg
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_natveg. Default values for the model are supplied (in metres).

layer_water
Vector of water (sf polygons) to be summarised. Coordinate reference system should be similar to points.

radii_water
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_water. Default values for the model are supplied (in metres).

layer_buildings
Vector of buildings (sf polygons) to be summarised. Coordinate reference system should be similar to points.

radii_buildings
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_buildings. Default values for the model are supplied (in metres).

buildings_levels
character. Specify column name for the number of levels within layer_buildings (if provided). Defaults to 'levels'. Column data should be numeric.

layer_roads
Vector of roads (sf lines) to be summarised. Coordinate reference system should be similar to points.

radii_roads
Numeric vector of radii (in the units of the projected coordinate reference system, e.g., metres) to summarise layer_roads. Default values for the model are supplied (in metres).

roads_lanes
character. Specify column name for the number of lanes within layer_roads (if provided). Defaults to 'lanes'. Column data should be numeric.

Value

The dataframe points with appended columns summarising the various landscape layers provided (at their respective radii).